+++ resources/views/admin/maintenance-records/show.blade.php
@@ -42,6 +42,10 @@
 
 @section('content')
 
+    @if (session('tire_size_warning'))
+        <div class="alert warning">{{ session('tire_size_warning') }}</div>
+    @endif
+
     <div class="page-actions">
         <a href="{{ request('back', route('admin.maintenance-records.index')) }}" class="btn ghost">
             <i class="fa-solid fa-arrow-left"></i> {{ __('Torna') }}